third_party/blink/web_tests/editing/shadow/delete-list-in-shadow-expected.txt

CONSOLE WARNING: line 22: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
Elements distributed to Shadow DOM should be deleted correctly when pressing delete key.

To test manually, select somewhere in ABCDE from somehwere in 12345, and press delete, and check the selected text is deleted correctly.

PASS hostTextContent is 'AB345'
PASS successfullyParsed is true

TEST COMPLETE
